Develop

+++ controllers/cards.js
@@ -1,11 +1,12 @@
 const Card = require('../models/card');
 const NotFoundError = require('../errors/notFoundError');
+const Forbidden = require('../errors/forbidden');
 
 const getCard = (req, res) => {
   Card.find()
     .orFail(() => new NotFoundError('Не удалось получить карточки'))
     .then((card) => res.send(card))
-    .catch((err) => res.status(err.statusCode || 500).send({ message: `${err.message}` }));
+    .catch((err) => res.status(err.statusCode || 500).send({ message: err.message }));
 };
 
 const createCard = (req, res) => {
@@ -16,9 +17,14 @@ const createCard = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
-    .orFail(() => new NotFoundError('Не удалось удалить карточку'))
-    .then((user) => res.send(user))
+  Card.findById(req.params.cardId).orFail(new NotFoundError('Не удалось найти карточку'))
+    .then((card) => {
+      if (!card.owner.equals(req.user._id)) {
+        throw new Forbidden('Вы не можете удалять чужие карточки');
+      }
+      return Card.findByIdAndDelete(req.params.cardId)
+        .then((cards) => res.send(cards));
+    })
     .catch((err) => res.status(err.statusCode || 500).send({ message: err.message }));
 };
 
@@ -30,7 +36,7 @@ const likesCard = (req, res) => {
   )
     .orFail(() => new NotFoundError('Не удалось поставить лайк'))
     .then((card) => res.send(card))
-    .catch((err) => res.status(err.statusCode || 500).send({ message: `${err.message}` }));
+    .catch((err) => res.status(err.statusCode || 500).send({ message: err.message }));
 };
 
 const deleteLike = (req, res) => {
@@ -41,7 +47,7 @@ const deleteLike = (req, res) => {
   )
     .orFail(() => new NotFoundError('Не удалось убрать лайк'))
     .then((card) => res.send(card))
-    .catch((err) => res.status(err.statusCode || 500).send({ message: `${err.message}` }));
+    .catch((err) => res.status(err.statusCode || 500).send({ message: err.message }));
 };
 
 module.exports = {
